Example archiving with Puppeteer and freeze-dry
Simple example script run-able from the command line. It will open a
visible instance of Puppeteer’s Chromium (for debugging) and load a new
tab with https://example.org/. It will then proceed to write both an
example.png and example.html file, containing a screenshot and
freeze-dry’d DOM respectively.

Change the URL on the first line of the file to test other pages.